Minimize number of files touched by C++ code generation and remove dead code
The class org.eclipse.xpand2.output.Outlet is used to set
file handling options for generated code.

The Outlet class has a setter for a 'VetoStrategy'.
The VetoStrategy::hasVeto(FileHandle handle) method is called
before the file will be written and after beforeWriteAndClose()
has been invoked on all post processors.

The special veto strategy:
'org.eclipse.xpand2.output.NoChangesVetoStrategy' was added.
It prevents writing of source code files without changes.

See also:

http://download.eclipse.org/modeling/m2t/xpand/javadoc/0.7.0/org/eclipse/xpand2/output/NoChangesVetoStrategy.html
http://heikobehrens.net/2009/04/23/generation-gap-pattern/

for further information.

The c++ source code generator got a new option '-c, --clear'.

Calling the generator with this option removes all files
in the package subdirectories, which are not related to an object
of the ecore model (i.e. which have not been generated in the
current run).