Override of default package factory and fix for typeid issue in class any
Creating a new (derived) model package factory will not delete
and reset the original _instance pointer of the base factory
any more. By this the replacement of the factory of the model
package will not fail anymore, because the old factory pointer
is still valid.

Recreation of Ecore model code and tests code.

Additional changes:

The type_id struct used in template class any doesn't
deliver a unique type id if the template is used in multiple
translation units. It was replaced by the save standard typeid()
function.