Parsing of pyecore library xmi files
The pyecore library uses the lxml library for serialization of
the xmi file. This library writes the XML declaration in the
following format:

<?xml version='1.0' encoding='UTF-8'?>
...

Strings are single quoted! This confuses the simple parser of the
emf4cpp library. The sematic rule for strings was extended to also
recognize single quoted strings.