EMF serializer shall support "keep default"
The load and save methods of class Resource got an additional string
to string map argument for overriding the default settings for parsing
and serialization of model instances.

Two new string constants are defined as static members in XMLResource.
These strings serve as key in the option map.

The XMLSerializer was extended to support the new options 'FORMATTED' and
'KEEP_DEFAULT_CONTENT'.

A new test was created in emf4cpp.tests/resource/test. The
ResourceOptions.ecore contains only one class ETypes with attributes
of the most common ETypes. For all of them a default value literal
is specified. In test-options.cpp an instance of this model is parsed
and serialized with and without option 'KEEP_DEFAULT_CONTENT' set.

The EType EChar is now mapped onto a char16_t in emf4cpp because
this is also the bit width of the corresponding java type char.
A new string_traits specialization for this type was created
and the generator was updated.

The generated code of all emf4cpp tests was updated to include
the recent changes of the generator.

Class any now dumps some type information in case of throwing
a bad_any_cast exception.