URIConverter for normalization of URLs
The new class URIConverter supports the mapping of URIs
and can provide input/output streams for certain URI types.

The mapping is done with the help of an internal string to
string map, which can be filled by the user.

When URIConverter::normalize(const QUrl& source) is called
the fragment and query are stripped from the source URL and
its string representation is looked up in the map. If a
mapping is found, the result gets 'normalized' again until
no further match is found in the map. The resulting target
URI is returned with fragment and query appended.

For providing output/input streams, check for the existence
of an URI resource or to remove an URI resource the URIConverter
delegates to a fitting URIHandler instance. To this end
the URIConverter holds a list of URIHandlers internally and
asks each handler in the list if it can handle the URI.
The first matching handler is used to satisfy the request.
If no handler is found an exception is thrown.

The new (base) class 'URIHandler' knows how to handle local
file URIs. For such URIs its URIHandler::canHandle() method
returns true.

The ResourceSet and the Resource uses the new URIConverter
for normalization of URIs in ResourceSet::getResource()
and for getting the right input/output stream in the
Resource::load()/save() methods.

The meta model in emf4cpp.tests/resource was extended and
renamed to 'ResourceTests.ecore'. A new test program
'test-uriconverter.cpp' for testing the use of alias
URIs in cross document references was added.

Additional changes:

- The ResourceContentEList is now a private nested class
  of 'Resource' to have access to the private data of
  the latter. When content is added to the Resource its
  state changes to 'Resource::isLoaded()==true'. Otherwise
  it would be loaded again, when a cross reference pointing
  to an object of this resource gets resolved.