Added support of enumerations and string type default literals
Enumeration attributes of classes were mapped to simple int
values during c++ code generation previously. Therefore there
was no safe way of setting this attribute to a particular value.
Additionally the initialization of the enum attribute to a
default value literal possibly defined in the ecore model did not
work.

To overcome these shortcomings a c++ header file containing the
definition of a c++11 enum class is now generated for every
enumeration type of the ecore model.
Classes with attributes of type enum include the corresponding
header file and have now a strongly typed enum attribute which
can by set by using the literals defined in the enum class.

A forward declaration for all enums of the model is
added to the 'modelname_forward.hpp' file in order to use
the enums as parameters or return values of class methods
(operations of the EClass definition).

The any_traits template used in the eGet() and eSet() methods
during parsing/serialization of c++ model instances were extended
to treat enum attributes correctly.

The initialization of string and enum attributes from the
default value literal defined in the ecore model was fixed
in the generator.

The new c++11 ISO standard was enabled in the library by adding
the '-std=c++11' option to the compile flags.

All (deprecated) uses of std::auto_ptr smart pointers were
replaced by a std::unique_ptr.

A new class template 'PackageDeleter' was added to the
ecorecpp part of the library. This deleter has an boolean
flag signaling if the smartpointer is the owner of the
object and the object is only deleted if the flag is true.

The unique_ptr holding the singleton instance of a package
is created with this custom deleter. The package has now
a second static access function to get a raw pointer to the
package instance: getInstanceAndRemoveOwnership().
This function sets the _owner flag of the deleter of the
unique_ptr to false and the object hold by this pointer will not
be deleted by the unique_ptrs destructor.

This new instance access function is used to add sub-packages
to the containing reference list of the main package, because
this list takes the ownership of the sub-packages.

A new test project 'enumeration.ecore' was added to emf4cpp.tests/
showing the treatment of enumerations and default value literals
for strings and enumerations.

The file 'type_traits.hpp' was renamed to 'type_definitions.hpp'
because it only contains typedefs and its old name conflicts with
a new (c++11) standard include header which is used in the 'any_traits'
class.

The source code generated from the EMF meta model and from all
test examples was updated with the new code generator.

Additional changes:

Mapping.ext:
  - set type mapping of EBigInteger, ELong and ELongObject to int64_t

A typo was fixed in the templates (references).

Fixed treatment of EEnum literals:
  - the generator used the loop counter as value for the eenum
    literal instead of the value stored in the EEnum definition,

  - the EEnum::getEEnumLiteral(int value) member function did not
    explicitly compare the given value with the values stored in
	the literal vector but returned the literal at index position
	'value' only.

The shebang '#! /bin/sh' was replaced by '#! /bin/bash' in
update-examples.sh and generator-release.sh because sh is
mapped to dash on certain systems (debian, ubuntu) but
certain bash extensions are used in the scripts.

EList.hpp/EListImpl.hpp:
  - removed return statement in void function

  - New access methods for the iterators of the underlying
    container were added to the EList class template.