EObject::as<T>() crashes if called on boost::intrusive_ptr<T>(nullptr)
The former EObject::as() template method is now a free template function
of two template arguments taking the object to be casted as argument.
Two versions of this function exists - one for naked c pointers and one
for boost::instrusive_ptr. The check for convertibility to EObject was
removed because, if the creation of an intrusive_ptr from the casted
object is not possible, the user will get a compile time error anyways.
Therefore is seems appropriate to use the simple rule: naked/smart
pointer in, naked/smart pointer out.

The definition of this method and several other using declarations
were moved from EObject.hpp to ecore_forward.hpp.

The instanceOf() template function now returns a bool if the types are
convertible. It was replaced by function as() at all places necessary.
It will be only created in ecore_forward.hpp and not in all forward
declaration files.

The explicit use of boost::intrusive_ptr was replaced by the alias
::ecore::Ptr defined in ecore:forward.hpp at several places.

Change-Id: I3a68411900751265a50f7289e67f8f4c88b137fd